requirements.txt
numpy>=1.20.0
pandas>=1.3.0
matplotlib>=3.4.0
seaborn>=0.11.0
scikit-learn>=1.0.0
openai>=1.0.0
python-dotenv>=0.19.0
jupyter>=1.0.0
streamlit>=1.20.0